+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 4.6.1 ##
+
+1.  UI Toolkit: Fix `ShowInInspector` can not correctly draw `Gradient`, `Curve` & `Hash128`
+2.  UI Toolkit: Improve the label layout in horizontal layout for Unity's native types (`Vector2` etc.)
+3.  UI Toolkit: Fix `ShowInInspector` might fail to draw some properties, give errors and block the rest drawing process
+
 ## 4.6.0 ##
 
 UI Toolkit: `ListDrawerSettings` add `string extraSearch`, `string overrideSearch` to allow custom search.